+++ __init__.py
@@ -5,12 +5,15 @@
 import typing as _t
 
 from .node_dict_from_text import StringConstructorDictFromText
+from .node_dict_preview import StringConstructorDictPreview
 
 NODE_CLASS_MAPPINGS: _t.Dict[str, type] = {
 	'StringConstructorDictFromText': StringConstructorDictFromText,
+	'StringConstructorDictPreview': StringConstructorDictPreview,
 }
 NODE_DISPLAY_NAME_MAPPINGS: _t.Dict[str, str] = {
-	'StringConstructorDictFromText': "Format-Dict from Text"
+	'StringConstructorDictFromText': "Format-Dict from Text",
+	'StringConstructorDictPreview': "Preview Format-Dict",
 }
 
 __all__ = ['NODE_CLASS_MAPPINGS', 'NODE_DISPLAY_NAME_MAPPINGS']